% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpinejs.R
\name{alpinejs}
\alias{alpinejs}
\alias{use_alpinejs}
\alias{html_dependency_alpinejs}
\title{Alpine.js}
\usage{
use_alpinejs(minified = TRUE)

html_dependency_alpinejs(minified = TRUE, cdn = NULL, version = NULL)
}
\arguments{
\item{minified}{Use the minified Alpine.js file? Default is \code{TRUE}.}

\item{cdn}{The URL to an online location for Alpine.js, typically a CDN.
Include the URL, without the CSS file, e.g. the full URL minus the
\code{alpine.js} or \code{alpine.min.js}.}

\item{version}{The version of the Alpine.js resource hosted at the URL in
\code{cdn}.}
}
\value{
An \code{htmltools::tagList()} with the Alpine.js dependencies, or an
\code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}.
}
\description{
Alpine.js is a rugged, minimal framework for composing JavaScript behavior in
your markup. .
}
\section{Functions}{
\itemize{
\item \code{use_alpinejs}: Adds Alpine.js dependencies to your document.

\item \code{html_dependency_alpinejs}: Returns an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} with the
alpinejs dependencies. Most users will want to use \code{use_alpinejs()}.
}}

\references{
\href{https://github.com/alpinejs/alpine/}{Alpine.js}
}
