% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tachyons.R
\name{style_colors}
\alias{style_colors}
\title{Add Tachyon-Inspired Utility Color Classes}
\usage{
style_colors(..., .replace_underscores = TRUE, .as_text = FALSE)
}
\arguments{
\item{...}{Named colors. The color names need to be acceptable as valid CSS
classes, in particular they cannot contain periods. Minimal validation is
performed on the color names, and no validation is performed on the
\href{https://developer.mozilla.org/en-US/docs/Web/CSS/color_value}{color syntax}.}

\item{.replace_underscores}{Replace underscores in argument names to turn
color names like \code{"hot_pink"} to \code{"hot-pink"}.}

\item{.as_text}{If \code{TRUE}, returns the CSS as a character string (does not
include the \verb{<style>} tag).}
}
\description{
Adds additional utility classes for colors. For each named color, CSS classes
are added for the foreground color  as \verb{.<color>}, background color as
\verb{.bg-<color>}, border color as \verb{.b--<color>}, and similar classes for hover
states for each that starts with the prefix \verb{.hover-} (i.e. \verb{.hover-<color>},
\verb{.hover-bg-<color}, and \verb{.hover-b--<color>}).
}
\seealso{
Other tachyons: 
\code{\link{style_tachyons}()},
\code{\link{tachyons}}
}
\concept{tachyons}
